영화 정보, 영화 감상 후기 페이지 댓글 기능 구현
영화 정보 : 영화 API 수정 필요. 댓글 기능은 구현하였으나 댓글 추천 기능 넣는 것 고려중
영화 감상 후기 : 댓글 조회, 작성, 수정, 삭제 기능 구현 완료

+ 영화 API 수정 필요(수정되는 API에서 영화 포스터 가져올 것)

+++ src/main/java/com/portfolio/demo/project/vo/CommentMovPagenationVO.java
@@ -16,7 +16,7 @@ public class CommentMovPagenationVO {
     private int startPage; // 시작페이지
     private int endPage; // 끝페이지
     private List<CommentMov> commentMovsList; // 한 화면에 보여줄 게시글 개수
-    private int totalCommentCnt; // 총 게시글 개수
+    private Long totalCommentCnt; // 총 게시글 개수
     private int commentsPerPage; // 페이지당 출력할 글 개수 (15)
     private int lastPage; // (총 개수에 따른) 마지막 페이지(전체 페이지 개수)
     private int start, end; // 시작인덱스, 끝 인덱스
@@ -25,7 +25,7 @@ public class CommentMovPagenationVO {
         this.currentPageNo = 1;
     }
 
-    public CommentMovPagenationVO(int totalCommentCnt, int currentPageNo, List<CommentMov> commentMovsList, int commentsPerPage, int start, int end) {
+    public CommentMovPagenationVO(Long totalCommentCnt, int currentPageNo, List<CommentMov> commentMovsList, int commentsPerPage, int start, int end) {
         this.totalCommentCnt = totalCommentCnt;
         this.currentPageNo = currentPageNo;
         this.commentMovsList = commentMovsList;
@@ -42,7 +42,7 @@ public class CommentMovPagenationVO {
         if (totalCommentCnt == 0) {
             totalPageCnt = 0;
         } else {
-            totalPageCnt = totalCommentCnt / commentsPerPage;
+            totalPageCnt =  (int)(totalCommentCnt / commentsPerPage);
             if (totalCommentCnt % commentsPerPage > 0) {
                 totalPageCnt++;
             }
